Removed ExcelTestRunner2.java and added support for running ExcelTestRunner in multi-file data mode when the "suiteFile" parameter contains a file name.
